+++ packages/iov-ethereum/src/testconfig.spec.ts
@@ -13,6 +13,7 @@ import {
 import { Encoding } from "@iov/encoding";
 
 import { Erc20Options } from "./erc20";
+import { EthereumConnectionOptions } from "./ethereumconnection";
 
 const { fromHex } = Encoding;
 
@@ -23,7 +24,7 @@ export interface Erc20TransferTest {
 export interface EthereumNetworkConfig {
   readonly env: string;
   readonly base: string;
-  readonly wsUrl: string;
+  readonly connectionOptions: EthereumConnectionOptions;
   readonly chainId: ChainId;
   readonly minHeight: number;
   readonly mnemonic: string;
@@ -51,7 +52,6 @@ export interface EthereumNetworkConfig {
   };
   readonly gasPrice: Amount;
   readonly gasLimit: Amount;
-  readonly waitForTx: number;
   readonly scraper:
     | {
         readonly apiUrl: string;
@@ -75,7 +75,11 @@ const env = process.env.ETHEREUM_NETWORK || "local";
 const local: EthereumNetworkConfig = {
   env: "local",
   base: "http://localhost:8545",
-  wsUrl: "ws://localhost:8545/ws",
+  connectionOptions: {
+    wsUrl: "ws://localhost:8545/ws",
+    // Low values to speedup test execution on the local ganache chain (using instant mine)
+    pollInterval: 0.1,
+  },
   chainId: "ethereum-eip155-5777" as ChainId,
   minHeight: 0, // ganache does not auto-generate a genesis block
   mnemonic: "oxygen fall sure lava energy veteran enroll frown question detail include maximum",
@@ -154,7 +158,6 @@ const local: EthereumNetworkConfig = {
     fractionalDigits: 18,
     tokenTicker: "ETH" as TokenTicker,
   },
-  waitForTx: 100, // by default, ganache will instantly mine a new block for every transaction
   scraper: undefined,
   expectedErrorMessages: {
     insufficientFunds: /sender doesn't have enough funds to send tx/i,
@@ -219,7 +222,9 @@ const local: EthereumNetworkConfig = {
 const ropsten: EthereumNetworkConfig = {
   env: "ropsten",
   base: "https://ropsten.infura.io/",
-  wsUrl: "wss://ropsten.infura.io/ws",
+  connectionOptions: {
+    wsUrl: "wss://ropsten.infura.io/ws",
+  },
   chainId: "ethereum-eip155-3" as ChainId,
   minHeight: 4284887,
   mnemonic: "oxygen fall sure lava energy veteran enroll frown question detail include maximum",
@@ -274,7 +279,6 @@ const ropsten: EthereumNetworkConfig = {
     fractionalDigits: 18,
     tokenTicker: "ETH" as TokenTicker,
   },
-  waitForTx: 4000,
   scraper: {
     apiUrl: "https://api-ropsten.etherscan.io/api",
     address: "0x0A65766695A712Af41B5cfECAaD217B1a11CB22A" as Address,
@@ -298,7 +302,9 @@ const ropsten: EthereumNetworkConfig = {
 const rinkeby: EthereumNetworkConfig = {
   env: "rinkeby",
   base: "https://rinkeby.infura.io",
-  wsUrl: "wss://rinkeby.infura.io/ws",
+  connectionOptions: {
+    wsUrl: "wss://rinkeby.infura.io/ws",
+  },
   chainId: "ethereum-eip155-4" as ChainId,
   minHeight: 3211058,
   mnemonic: "retire bench island cushion panther noodle cactus keep danger assault home letter",
@@ -373,7 +379,6 @@ const rinkeby: EthereumNetworkConfig = {
     fractionalDigits: 18,
     tokenTicker: "ETH" as TokenTicker,
   },
-  waitForTx: 4000,
   scraper: {
     apiUrl: "https://api-rinkeby.etherscan.io/api",
     // recipient address with no known keypair